HW-03 power half

+++ 03_algebra/power_test.cpp
@@ -9,7 +9,7 @@ namespace algorithms_202104
 bool PowerTest::test_passed(std::ifstream & ifs_in, std::ifstream & ifs_out)
 {
     double A;
-    int N;
+    unsigned long N;
     ifs_in >> A >> N;
 
     //std::cout << A << " " << N << std::endl;
@@ -20,13 +20,12 @@ bool PowerTest::test_passed(std::ifstream & ifs_in, std::ifstream & ifs_out)
     //std::cout << std::setprecision(15) << answer << std::endl;
 
     double rezult = 1.0;
-    for (int k = 0; k < N; ++k)
+    for (unsigned long k = 0; k < N; ++k)
     {
         rezult *= A;
     }
 
-    //std::cout << rezult << std::endl;
-
+    //std::cout  << std::setprecision(14) << rezult << std::endl;
 
     return std::fabs(answer - rezult) < 1e-10;
 }
